+++ NEWS.md
@@ -1,6 +1,7 @@
 # Version 0.0.8 11-20-2018
   * rewrote and consolidate code for fluglm
   * updated tidy eval for most recent version of rlang
+  * changed fluglm output to be predicted and baseline fitted values
   
 
 # Version 0.0.7 08-08-2018